package/wpewebkit: bump to version 2.30.1
Update to a new major release which bring in many improvements and new
features. The .0 version is skipped because it includes a regression
which was fixed by the .1 release. For a complete list of changed,
please refer to the release notes:

  https://wpewebkit.org/release/wpewebkit-2.30.0.html
  https://wpewebkit.org/release/wpewebkit-2.30.1.html

None of the new features in WPE WebKit 2.30.x need additional dependencies.

Signed-off-by: Adrian Perez de Castro <aperez@igalia.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>