package/mosquitto: bump version to 2.0.0
A new major version, see the announcement for details:
https://mosquitto.org/blog/2020/12/version-2-0-0-released/

License has now changed to v2.0 of the Eclipse Public License, so update the
license info and hashes to match.

There is now optional cJSON support, so handle that.

Add upstream post-2.0.0 patches fixing build with cJSON and without TLS
support.

Signed-off-by: Peter Korsgaard <peter@korsgaard.com>