package/ti-sgx-{km, um}: bump to SDK 06.01.00.08 versions
Currently, the ti-sgx packages and the beaglebone_qt5_defconfig do
not work with KMS nor Weston. What's worse, is the latest SDK version
06.03.00.106 (as of this commit) of these packages is broken and does
not correctly support KMS, and attempting to run KMS applications
results in eglfs initialization failures. As such, bumping these
packages to the version before 06.03.00.106 is the best option.

Because of the above problems, several packages must change at the
same time to ensure this patch does not break any other packages:

  - ti-sgx-libgbm
    - dropped, merged into ti-sgx-um, see below

  - ti-sgx-um:
    - bump the version that matches TI SDK 06.01.00.08.
    - demove select BR2_PACKAGE_TI_SGX_LIBGBM in Config.in, as the libgbm
      package merges ti-sgx-libgbm with this package.

  - ti-sgx-km:
    - bump the version that matches TI SDK 06.01.00.08.

Signed-off-by: Adam Duskett <Aduskett@gmail.com>
Tested-by: Markus <zehnder@live.com>
[yann.morin.1998@free.fr:
  - actually switch qt5base to use ti-sgx-um
  - split the beaglebone config changes to their own patch
  - split the ti-sgx-demos changes to their own patch
]
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>